// frontend/src/declarations/icp_ledger/icp_ledger.did.d.ts
export type BlockIndex = bigint;
export interface BlockRange {
	blocks: Array<Block>;
}
export interface Duration {
	secs: bigint;
	nanos: number;
}
export interface FeatureFlags {
	icrc2: boolean;
}
export interface GetBlocksArgs {
	start: BlockIndex;
	length: bigint;
}
export type Icrc1BlockIndex = bigint;
export type Icrc1Timestamp = bigint;
export type Icrc1Tokens = bigint;
export type Icrc1TransferError =
	| {
			GenericError: { message: string; error_code: bigint };
	  }
	| { TemporarilyUnavailable: null }
	| { BadBurn: { min_burn_amount: Icrc1Tokens } }
	| { Duplicate: { duplicate_of: Icrc1BlockIndex } }
	| { BadFee: { expected_fee: Icrc1Tokens } }
	| { CreatedInFuture: { ledger_time: bigint } }
	| { TooOld: null }
	| { InsufficientFunds: { balance: Icrc1Tokens } };
// ...
export interface Tokens {
	e8s: bigint;
}
// ...
export type TransferError =
	| {
			TxTooOld: { allowed_window_nanos: bigint };
	  }
	| { BadFee: { expected_fee: Tokens } }
	| { TxDuplicate: { duplicate_of: BlockIndex } }
	| { TxCreatedInFuture: null }
	| { InsufficientFunds: { balance: Tokens } };